unify matmuls
